requirements.txt
Pillow
imageio 